abstract ModalSpace out of Modal.tsx
